+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Version 1.4.0
 - Updating config appearances whenever a `ConfigEntry` is modified directly.
+- Added `QueueCustomConfigFileForAutoGeneration` method, allowing the auto generation of configs for manually created config files.
 
 ## Version 1.3.4
 - Fixed issue where some assemblies' depedencies could cause the UI to not load properly.